Implement database access logic
