+++ src/Infrastructure/Data/DbInitializer.cs
@@ -1,0 +1,30 @@
+﻿using ApplicationCore.Entities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Infrastructure.Data;
+
+public static class DbInitializer
+{
+    private static readonly string[] Summaries = new[]
+    {
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    };
+
+    public static async Task SeedAsync(IServiceProvider serviceProvider)
+    {
+        using (var context = new DatabaseContext(serviceProvider.GetRequiredService<DbContextOptions<DatabaseContext>>()))
+        {
+            context.Database.EnsureCreated();
+
+            var entities = Enumerable.Range(1, 5).Select(index => new WeatherForecast
+            {
+                Date = DateTime.Now.AddDays(index),
+                TemperatureC = Random.Shared.Next(-20, 55),
+                Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+            });
+            await context.WeatherForecasts.AddRangeAsync(entities);
+            await context.SaveChangesAsync();
+        }
+    }
+}
