+++ src/Infrastructure/DatabaseContext.cs
@@ -7,6 +7,7 @@ public class DatabaseContext : DbContext
 {
     public DatabaseContext(DbContextOptions<DatabaseContext> options) : base(options)
     {
+        AppContext.SetSwitch("Npgsql.EnableLegacyTimestampBehavior", true);
     }
 
     public DbSet<WeatherForecast> WeatherForecasts { get; set; } = null!;
