+++ src/Infrastructure/WeatherForecastRepository.cs
@@ -16,20 +16,10 @@ public class WeatherForecastRepository : IWeatherForecastRepository
         _context = context;
     }
 
-    private static readonly string[] Summaries = new[]
-    {
-        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-    };
-
     /// <inheritdoc/>
     public IEnumerable<WeatherForecast> GetWeatherForecasts()
     {
-        var entities = Enumerable.Range(1, 5).Select(index => new WeatherForecast
-        {
-            Date = DateTime.Now.AddDays(index),
-            TemperatureC = Random.Shared.Next(-20, 55),
-            Summary = Summaries[Random.Shared.Next(Summaries.Length)]
-        });
+        var entities = _context.WeatherForecasts.AsEnumerable();
         return entities;
     }
 }
