+++ src/WebApi/Program.cs
@@ -4,6 +4,7 @@ using ApplicationCore.Interfaces;
 using ApplicationCore.Mapper;
 using ApplicationCore.Services;
 using Infrastructure;
+using Infrastructure.Data;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.OpenApi.Models;
@@ -85,5 +86,12 @@ app.UseHttpsRedirection();
 app.UseAuthorization();
 
 app.MapControllers();
+
+// Database initialization process
+using (var scope = app.Services.CreateScope())
+{
+    var services = scope.ServiceProvider;
+    await DbInitializer.SeedAsync(services);
+}
 
 app.Run();
